+++ packages/firestore/exp/index.ts
@@ -15,10 +15,9 @@
  * limitations under the License.
  */
 
-import { version } from '../package.json';
-import { _registerComponent, registerVersion } from '@firebase/app-exp';
-import { Component, ComponentType } from '@firebase/component';
-import { Firestore } from './src/api/database';
+import { registerFirestore } from './register';
+
+registerFirestore();
 
 export { FieldPath, documentId } from '../lite/src/api/field_path';
 
@@ -71,9 +70,9 @@ export {
   getDoc,
   getDocFromCache,
   getDocFromServer,
-  getQuery,
-  getQueryFromCache,
-  getQueryFromServer,
+  getDocs,
+  getDocsFromCache,
+  getDocsFromServer,
   onSnapshot,
   onSnapshotsInSync,
   setDoc,
@@ -104,22 +103,3 @@ export { GeoPoint } from '../src/api/geo_point';
 export { Timestamp } from '../src/api/timestamp';
 
 export { refEqual, queryEqual } from '../lite/src/api/reference';
-
-export function registerFirestore(): void {
-  _registerComponent(
-    new Component(
-      'firestore-exp',
-      container => {
-        const app = container.getProvider('app-exp').getImmediate()!;
-        return ((app, auth) => new Firestore(app, auth))(
-          app,
-          container.getProvider('auth-internal')
-        );
-      },
-      ComponentType.PUBLIC
-    )
-  );
-  registerVersion('firestore-exp', version, 'node');
-}
-
-registerFirestore();